finish pull request comments

+++ src/main/java/io/codelex/advanced/exercise1/Main.java
@@ -1,13 +1,15 @@
 package io.codelex.advanced.exercise1;
 
+import java.math.BigDecimal;
+
 public class Main {
     public static void main(String[] args) throws NotEnoughFundsException {
-        DebitCard debitCard = new DebitCard(4144, "Joe Holmes", 113, 9999);
-        CreditCard creditCard = new CreditCard(5551, "Jimmy Bob", 561, 120);
+        DebitCard debitCard = new DebitCard(4144, "Joe Holmes", 113, BigDecimal.valueOf(9999));
+        CreditCard creditCard = new CreditCard(5551, "Jimmy Bob", 561, BigDecimal.valueOf(120));
 
         //prints a warning
-        debitCard.addMoney(12);
+        debitCard.addMoney(BigDecimal.valueOf(12));
         //throws exception
-        creditCard.withdrawMoney(150);
+        creditCard.withdrawMoney(BigDecimal.valueOf(150));
     }
 }
